add automatic retries for failed ticket validations
This handles the unfortunately-not-too-uncommon INVALID_TICKET error more gracefully.
One cause of this error is ticket timeouts. Currently, the cas server is configured to allow tickets to live for 30 seconds before becoming invalid. It appears that in some cases (though it's hard to imagine why) the validation step is occurring more than 30 seconds after the granting of the service ticket.
There may be other causes for INVALID_TICKET. Hopefully, they will be ephemeral, and a obtaining a new service ticket will solve the problem. In case it doesn't, we will stop re-acquiring service tickets after 3 attempts.