add ability to store 'logout list' in infinispan
This abstracts the storage of logout requests, and makes it possible to store them in a distributed cache (infinispan) instead of a static in-jvm list.
With this, the LogoutEchoFilter is not needed. This allows docker-based apps to correctly handle logout requests, because each node will not have a static hostname.